app/requirements.txt
python-dotenv==0.19.2
psycopg2-binary==2.9.2
confluent-kafka==1.7.0